requirements.txt
Flask==1.1.2
Flask-Caching==1.5.0
pyjq==2.4.0
PyYAML==5.3.1
requests==2.23.0
envsubst==0.1.5
gunicorn==20.0.4
six==1.11.0
